pyproject.toml
[tool.poetry]
name = "stockscreener"
version = "0.1.0"
description = "Flask-based stock screener"
authors = ["fayssal el mofatiche <fayssal.elmofatiche@dnbf.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-login = "^0.6.3"
email-validator = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
